#3 reorganizes structure so not everything is inside App component
