+++ src/components/game-arena/GameArena.test.js
@@ -1,17 +1,11 @@
 import React from "react";
-import App from "./App";
-import { shallow, mount } from "enzyme";
+import { GameArena } from "./GameArena";
+import { mount } from "enzyme";
 import "isomorphic-fetch";
 
-describe("App Component", () => {
-  it('displays "Picture Naming Game" as the title', () => {
-    const wrapper = shallow(<App />);
-    const text = wrapper.find("h1").text();
-    expect(text).toEqual("Picture Naming Game");
-  });
-
+describe("GameArena Component", () => {
   it("does not display an imageCanvas unless there is an image fetched", () => {
-    const wrapper = mount(<App />);
+    const wrapper = mount(<GameArena />);
     let imageCanvas = wrapper.find("ImageCanvas");
     expect(imageCanvas.length).toBe(0);
     wrapper.setState({
