+++ src/components/game-arena/image-canvas/ImageCanvas.js
