+++ src/components/game-arena/image-canvas/ImageCanvas.test.js
